requirements.txt
anthropic
flask
playwright
python-dotenv
openpyxl
PyPDF2
python-docx
flask-limiter
requests
google-api-python-client>=2.100.0
google-auth>=2.26.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
boto3>=1.28.0
